Airdrop API and snoarqube related cleanup
