+++ src/main/java/com/airdropmc/PackagesConfig.java
@@ -1,18 +1,15 @@
-package lukemccon.airdrop;
+package com.airdropmc;
 
 import java.io.File;
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
 
+import com.airdropmc.helpers.ChatHandler;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.configuration.file.YamlConfiguration;
 import org.bukkit.inventory.ItemStack;
-
-import lukemccon.airdrop.helpers.ChatHandler;
-
-import static lukemccon.airdrop.Airdrop.PLUGIN_NAME;
 
 public class PackagesConfig {
 
@@ -25,12 +22,12 @@ public class PackagesConfig {
 	private static boolean loaded = false;
 	
 	public static void loadConfig() {
-		f = new File(Bukkit.getServer().getPluginManager().getPlugin(PLUGIN_NAME).getDataFolder(), PACKAGES_FILE_NAME);
+		f = new File(Bukkit.getServer().getPluginManager().getPlugin(Airdrop.PLUGIN_NAME).getDataFolder(), PACKAGES_FILE_NAME);
 		if(f.exists()) {
 			yaml = YamlConfiguration.loadConfiguration(f);
 			loaded = true;
 		} else {
-			f = new File(Bukkit.getServer().getPluginManager().getPlugin(PLUGIN_NAME).getDataFolder(), PACKAGES_FILE_NAME);
+			f = new File(Bukkit.getServer().getPluginManager().getPlugin(Airdrop.PLUGIN_NAME).getDataFolder(), PACKAGES_FILE_NAME);
 			yaml = YamlConfiguration.loadConfiguration(f);
 			try {
 				yaml.save(f);
@@ -44,7 +41,7 @@ public class PackagesConfig {
 	}
 
 	public static void saveConfig(YamlConfiguration config) {
-		f = new File(Bukkit.getServer().getPluginManager().getPlugin(PLUGIN_NAME).getDataFolder(), PACKAGES_FILE_NAME);
+		f = new File(Bukkit.getServer().getPluginManager().getPlugin(Airdrop.PLUGIN_NAME).getDataFolder(), PACKAGES_FILE_NAME);
 		try {
 			config.save(f);
 		} catch (IOException e) {
