+++ src/main/java/com/airdropmc/controllers/PackageController.java
@@ -1,49 +1,26 @@
-package lukemccon.airdrop.controllers;
+package com.airdropmc.controllers;
 
-import lukemccon.airdrop.exceptions.PackageNotFoundException;
-import lukemccon.airdrop.helpers.ChatHandler;
-import lukemccon.airdrop.packages.CreatePackageGui;
-import lukemccon.airdrop.packages.PackageManager;
+import com.airdropmc.helpers.ChatHandler;
+import com.airdropmc.packages.CreatePackageGui;
+import com.airdropmc.packages.PackageManager;
+import com.airdropmc.exceptions.PackageNotFoundException;
 import org.bukkit.ChatColor;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 
-import java.util.Objects;
-
 public class PackageController {
 
-	PackageController() {
+	private PackageController() {
 
 	}
-	
-	public static boolean onCommand(CommandSender sender, String[] args) {
 
-		if (args.length < 2) {
-			ChatHandler.sendErrorMessage(sender, "Must specify a package name");
-			ChatHandler.sendErrorMessage(sender, "Example: /airdrop starter");
-			return true;
-		}
 
-		if (Objects.equals(args[1], "create")) {
-			return createPackageCommand(sender, args);
-		}
-
-		if (Objects.equals(args[1], "delete")) {
-			return deletePackageCommand(sender, args);
-		}
-
-		String packageName = args[1];
-
-		try {
-			ChatHandler.sendMessage(sender, ChatColor.WHITE + "\nPackage info for: " + ChatColor.AQUA + packageName + "\n================\n" +  ChatColor.WHITE + PackageManager.getInfo(packageName));
-		} catch (PackageNotFoundException e) {
-			ChatHandler.sendErrorMessage(sender, e.getMessage());
-		}
-
-		return true;
-	}
-
-	public static boolean deletePackageCommand(CommandSender sender, String[] args) {
+	/**
+	 * Subcommand that handles package deletion
+	 * @param sender who executed the command
+	 * @param args provided to the command
+	 */
+	public static void deletePackageCommand(CommandSender sender, String[] args) {
 		if (args.length != 3 ) {
 			ChatHandler.sendErrorMessage(sender, "Need to specify a package name to delete");
 		}
@@ -54,12 +31,15 @@ public class PackageController {
 		} catch (PackageNotFoundException e) {
 			ChatHandler.sendErrorMessage(sender, "Unable to delete package: " + ChatColor.DARK_RED + packageName + ChatColor.RED + " not found");
 		}
-		return true;
 	}
 
 
-
-	public static boolean createPackageCommand(CommandSender sender, String[] args) {
+	/**
+	 * Subcommand that handles package creation
+	 * @param sender of command
+	 * @param args command args
+	 */
+	public static void createPackageCommand(CommandSender sender, String[] args) {
 
 			if (args.length != 4) {
 				ChatHandler.sendErrorMessage(sender, "Package create command requires 4 total arguments");
@@ -67,20 +47,18 @@ public class PackageController {
 			}
 
 			// Create a new package
-			if (!(sender instanceof Player)) {
+			if (!(sender instanceof Player player)) {
 				ChatHandler.sendErrorMessage(sender,"Must be a player to use this command");
-				return true;
+				return;
 			}
 
-			Player player = (Player) sender;
-
-			String packageName = args[2];
+		String packageName = args[2];
 			String priceString = args[3];
-			Double price = null;
+			double price = 0;
 
 			if ( packageName == null || packageName.isBlank()) {
 				ChatHandler.sendErrorMessage(sender, "You must provide a name for the package");
-				return true;
+				return;
 			}
 
 			if ( priceString != null && !priceString.isBlank()) {
@@ -89,7 +67,7 @@ public class PackageController {
 				} catch (NumberFormatException e) {
 					ChatHandler.sendErrorMessage(sender, "You must provide the package price as a double");
 					ChatHandler.sendErrorMessage(sender, "Example: /airdrop package create myPackage 12.0");
-					return true;
+					return;
 				}
 			}
 
@@ -97,9 +75,6 @@ public class PackageController {
 			CreatePackageGui createGui = new CreatePackageGui(packageName, price);
 
 			createGui.openInventory(player);
-
-			return true;
-
 	}
 
 }
