+++ src/main/java/com/airdropmc/exceptions/CannotAffordException.java
@@ -1,0 +1,8 @@
+package com.airdropmc.exceptions;
+
+public class CannotAffordException extends Exception {
+
+    public CannotAffordException(String message){
+        super(message);
+    }
+}
