+++ src/main/java/com/airdropmc/exceptions/InsufficientPermissionsException.java
@@ -1,0 +1,10 @@
+package com.airdropmc.exceptions;
+
+public class InsufficientPermissionsException extends Exception {
+    /**
+     * Indicates that a player does not have permissions to carry out an operation
+     */
+    public InsufficientPermissionsException(String message) {
+        super(message);
+    }
+}
