+++ src/main/java/com/airdropmc/listeners/FallingBlockListener.java
@@ -1,4 +1,4 @@
-package lukemccon.airdrop.listeners;
+package com.airdropmc.listeners;
 
 import org.bukkit.entity.Entity;
 import org.bukkit.event.EventHandler;
@@ -6,8 +6,8 @@ import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
 import org.bukkit.event.entity.EntityChangeBlockEvent;
 
-import lukemccon.airdrop.helpers.CrateList;
-import lukemccon.airdrop.Crate;
+import com.airdropmc.helpers.CrateList;
+import com.airdropmc.Crate;
 
 public class FallingBlockListener implements Listener {
